requirements.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.1.1
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.5.0
asttokens==2.4.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
autopep8==2.0.2
Babel==2.12.1
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
billiard==4.1.0
bleach==6.0.0
blinker==1.5
branca==0.6.0
celery==5.3.4
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.1.4
cryptography==38.0.4
DateTime==5.2
debugpy==1.8.0
decorator==5.1.1
decouple==0.0.7
defusedxml==0.7.1
Django==4.0.3
django-cors-headers==3.11.0
django-dotenv==1.4.2
django-filter==22.1
django-multiselectfield==0.1.12
django-smart-selects==1.6.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
executing==1.2.0
folium==0.14.0
fqdn==1.5.1
frozenlist==1.4.0
greenlet==2.0.1
httpie==3.2.1
idna==3.4
ipykernel==6.25.2
ipython==8.15.0
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kombu==5.3.2
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1
multidict==6.0.4
mysql==0.0.3
mysql-connector-python==8.0.31
mysqlclient==2.1.1
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.7
notebook==7.0.3
notebook_shim==0.2.3
numpy==1.23.5
opencage==2.1.0
opencv-python==4.6.0.66
overrides==7.4.0
packaging==23.1
pandocfilters==1.5.0
parso==0.8.3
phonenumbers==8.13.2
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.10.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==3.20.1
psutil==5.9.5
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
Pygments==2.14.0
PyJWT==2.3.0
PyMySQL==1.0.2
pyOpenSSL==22.1.0
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
python-json-logger==2.0.7
pytz==2021.3
pywin32==306
pywinpty==2.0.11
PyYAML==6.0
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
redis==5.0.1
referencing==0.30.2
requests==2.31.0
requests-toolbelt==0.10.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.2
rpds-py==0.10.3
scipy==1.9.3
Send2Trash==1.8.2
six==1.16.0
sketchpy==0.1.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==1.4.45
sqlparse==0.4.2
stack-data==0.6.2
svg.path==6.2
svgpathtools==1.5.1
svgwrite==1.4.3
terminado==0.17.1
tinycss2==1.2.1
tk==0.1.0
tornado==6.3.3
tqdm==4.64.1
traitlets==5.10.0
turtle==0.0.1
twilio==8.10.0
tzdata==2023.3
uri-template==1.3.0
urllib3==2.0.7
vine==5.0.0
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
Werkzeug==2.2.2
widgetsnbextension==4.0.9
yarl==1.9.2
zope.interface==6.0
